requirement.txt
sydney-py
discord.py-self
